safari-534.53-branch/LayoutTests/platform/mac/fast/dom/objc-wrapper-identity-expected.txt

This tests wrapper identity for Objective-C.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS var object = [ ]; isEqualObjC(object, object); is true
PASS var object = { }; isEqualObjC(object, object); is true
PASS var object = new Array(); isEqualObjC(object, object); is true
PASS var object = new RegExp(); isEqualObjC(object, object); is true
PASS var object = new String(); isEqualObjC(object, object); is true
PASS var object = window['CSSPrimitiveValue']; isEqualObjC(object, object); is true
PASS var object = window['CSSRule']; isEqualObjC(object, object); is true
PASS var object = window['CSSStyleDeclaration']; isEqualObjC(object, object); is true
PASS var object = window['CSSValue']; isEqualObjC(object, object); is true
PASS var object = window['CSSCharsetRule']; isEqualObjC(object, object); is true
PASS var object = window['CSSImportRule']; isEqualObjC(object, object); is true
PASS var object = window['CSSMediaRule']; isEqualObjC(object, object); is true
PASS var object = window['CSSFontFaceRule']; isEqualObjC(object, object); is true
PASS var object = window['CSSPageRule']; isEqualObjC(object, object); is true
PASS var object = window['CSSVariablesRule']; isEqualObjC(object, object); is true
PASS var object = window['WebKitCSSKeyframeRule']; isEqualObjC(object, object); is true
PASS var object = window['WebKitCSSKeyframesRule']; isEqualObjC(object, object); is true
PASS var object = window['DOMException']; isEqualObjC(object, object); is true
PASS var object = window['DOMParser']; isEqualObjC(object, object); is true
PASS var object = window['Document']; isEqualObjC(object, object); is true
PASS var object = window['DocumentFragment']; isEqualObjC(object, object); is true
PASS var object = window['Element']; isEqualObjC(object, object); is true
PASS var object = window['EvalError']; isEqualObjC(object, object); is true
PASS var object = window['Event']; isEqualObjC(object, object); is true
PASS var object = window['HTMLAnchorElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLAppletElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLAreaElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLBRElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLBaseElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLBaseFontElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLBodyElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLButtonElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLCanvasElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLDListElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLDirectoryElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLDivElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLDocument']; isEqualObjC(object, object); is true
PASS var object = window['HTMLElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLFieldSetElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLFontElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLFormElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLHRElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLHeadElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLHeadingElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLHtmlElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLImageElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLInputElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLIsIndexElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLLIElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLLabelElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLLegendElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLLinkElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLMapElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLMenuElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLMetaElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLModElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLOListElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLOptGroupElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLOptionElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLParagraphElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLParamElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLPreElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLQuoteElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLScriptElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLSelectElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLStyleElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLTextAreaElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLTitleElement']; isEqualObjC(object, object); is true
PASS var object = window['HTMLUListElement']; isEqualObjC(object, object); is true
PASS var object = window['Image']; isEqualObjC(object, object); is true
PASS var object = window['MutationEvent']; isEqualObjC(object, object); is true
PASS var object = window['Node']; isEqualObjC(object, object); is true
PASS var object = window['NodeFilter']; isEqualObjC(object, object); is true
PASS var object = window['Option']; isEqualObjC(object, object); is true
PASS var object = window['Range']; isEqualObjC(object, object); is true
PASS var object = window['RangeError']; isEqualObjC(object, object); is true
PASS var object = window['RangeException']; isEqualObjC(object, object); is true
PASS var object = window['ReferenceError']; isEqualObjC(object, object); is true
PASS var object = window['SyntaxError']; isEqualObjC(object, object); is true
PASS var object = window['TypeError']; isEqualObjC(object, object); is true
PASS var object = window['URIError']; isEqualObjC(object, object); is true
PASS var object = window['XMLDocument']; isEqualObjC(object, object); is true
PASS var object = window['XMLHttpRequest']; isEqualObjC(object, object); is true
PASS var object = window['XMLSerializer']; isEqualObjC(object, object); is true
PASS var object = window['XPathEvaluator']; isEqualObjC(object, object); is true
PASS var object = window['XPathResult']; isEqualObjC(object, object); is true
PASS var object = window['XSLTProcessor']; isEqualObjC(object, object); is true
PASS var object = window['addEventListener']; isEqualObjC(object, object); is true
PASS var object = window['alert']; isEqualObjC(object, object); is true
PASS var object = window['atob']; isEqualObjC(object, object); is true
PASS var object = window['blur']; isEqualObjC(object, object); is true
PASS var object = window['btoa']; isEqualObjC(object, object); is true
PASS var object = window['captureEvents']; isEqualObjC(object, object); is true
PASS var object = window['clearInterval']; isEqualObjC(object, object); is true
PASS var object = window['clearTimeout']; isEqualObjC(object, object); is true
PASS var object = window['clientInformation']; isEqualObjC(object, object); is true
PASS var object = window['close']; isEqualObjC(object, object); is true
PASS var object = window['closed']; isEqualObjC(object, object); is true
PASS var object = window['confirm']; isEqualObjC(object, object); is true
PASS var object = window['console']; isEqualObjC(object, object); is true
PASS var object = window['crypto']; isEqualObjC(object, object); is true
PASS var object = window['defaultStatus']; isEqualObjC(object, object); is true
PASS var object = window['defaultstatus']; isEqualObjC(object, object); is true
PASS var object = window['document']; isEqualObjC(object, object); is true
PASS var object = window['event']; isEqualObjC(object, object); is true
PASS var object = window['find']; isEqualObjC(object, object); is true
PASS var object = window['focus']; isEqualObjC(object, object); is true
PASS var object = window['frameElement']; isEqualObjC(object, object); is true
PASS var object = window['frames']; isEqualObjC(object, object); is true
PASS var object = window['getComputedStyle']; isEqualObjC(object, object); is true
PASS var object = window['getMatchedCSSRules']; isEqualObjC(object, object); is true
PASS var object = window['getSelection']; isEqualObjC(object, object); is true
PASS var object = window['history']; isEqualObjC(object, object); is true
PASS var object = window['location']; isEqualObjC(object, object); is true
PASS var object = window['locationbar']; isEqualObjC(object, object); is true
PASS var object = window['log']; isEqualObjC(object, object); is true
PASS var object = window['menubar']; isEqualObjC(object, object); is true
PASS var object = window['moveBy']; isEqualObjC(object, object); is true
PASS var object = window['moveTo']; isEqualObjC(object, object); is true
PASS var object = window['name']; isEqualObjC(object, object); is true
PASS var object = window['navigator']; isEqualObjC(object, object); is true
PASS var object = window['offscreenBuffering']; isEqualObjC(object, object); is true
PASS var object = window['onabort']; isEqualObjC(object, object); is true
PASS var object = window['onbeforeunload']; isEqualObjC(object, object); is true
PASS var object = window['onblur']; isEqualObjC(object, object); is true
PASS var object = window['onchange']; isEqualObjC(object, object); is true
PASS var object = window['onclick']; isEqualObjC(object, object); is true
PASS var object = window['ondblclick']; isEqualObjC(object, object); is true
PASS var object = window['onerror']; isEqualObjC(object, object); is true
PASS var object = window['onfocus']; isEqualObjC(object, object); is true
PASS var object = window['onkeydown']; isEqualObjC(object, object); is true
PASS var object = window['onkeypress']; isEqualObjC(object, object); is true
PASS var object = window['onkeyup']; isEqualObjC(object, object); is true
PASS var object = window['onload']; isEqualObjC(object, object); is true
PASS var object = window['onmousedown']; isEqualObjC(object, object); is true
PASS var object = window['onmousemove']; isEqualObjC(object, object); is true
PASS var object = window['onmouseout']; isEqualObjC(object, object); is true
PASS var object = window['onmouseover']; isEqualObjC(object, object); is true
PASS var object = window['onmouseup']; isEqualObjC(object, object); is true
PASS var object = window['onmousewheel']; isEqualObjC(object, object); is true
PASS var object = window['onreset']; isEqualObjC(object, object); is true
PASS var object = window['onresize']; isEqualObjC(object, object); is true
PASS var object = window['onscroll']; isEqualObjC(object, object); is true
PASS var object = window['onsearch']; isEqualObjC(object, object); is true
PASS var object = window['onselect']; isEqualObjC(object, object); is true
PASS var object = window['onsubmit']; isEqualObjC(object, object); is true
PASS var object = window['onunload']; isEqualObjC(object, object); is true
PASS var object = window['open']; isEqualObjC(object, object); is true
PASS var object = window['opener']; isEqualObjC(object, object); is true
PASS var object = window['parent']; isEqualObjC(object, object); is true
PASS var object = window['personalbar']; isEqualObjC(object, object); is true
PASS var object = window['print']; isEqualObjC(object, object); is true
PASS var object = window['prompt']; isEqualObjC(object, object); is true
PASS var object = window['releaseEvents']; isEqualObjC(object, object); is true
PASS var object = window['removeEventListener']; isEqualObjC(object, object); is true
PASS var object = window['resizeBy']; isEqualObjC(object, object); is true
PASS var object = window['resizeTo']; isEqualObjC(object, object); is true
PASS var object = window['screen']; isEqualObjC(object, object); is true
PASS var object = window['scroll']; isEqualObjC(object, object); is true
PASS var object = window['scrollBy']; isEqualObjC(object, object); is true
PASS var object = window['scrollTo']; isEqualObjC(object, object); is true
PASS var object = window['scrollbars']; isEqualObjC(object, object); is true
PASS var object = window['self']; isEqualObjC(object, object); is true
PASS var object = window['setInterval']; isEqualObjC(object, object); is true
PASS var object = window['setTimeout']; isEqualObjC(object, object); is true
PASS var object = window['showModalDialog']; isEqualObjC(object, object); is true
PASS var object = window['status']; isEqualObjC(object, object); is true
PASS var object = window['statusbar']; isEqualObjC(object, object); is true
PASS var object = window['stop']; isEqualObjC(object, object); is true
PASS var object = window['test']; isEqualObjC(object, object); is true
PASS var object = window['toolbar']; isEqualObjC(object, object); is true
PASS var object = window['top']; isEqualObjC(object, object); is true
PASS var object = window['window']; isEqualObjC(object, object); is true
